Moved homePage and mapTest composable to views directory. Created model directory with Country.kt file.
