+++ app/src/main/java/com/example/capstoneapp/model/Country.kt
@@ -1,0 +1,6 @@
+package com.example.capstoneapp.model
+
+class Country {
+
+}
+
